+++ CHANGELOG.md
@@ -1,11 +1,16 @@
-### SymbolicRegressionToolkit-1.1.1 (TBA)
+### SymbolicRegressionToolkit-1.2.0 (TBA)
 
 - Updated documentation
 - Added the change log
 
 **New Features:**
 
-- Added Datasets
+- Added Dataset and benchmark objects that create evaluators for Symbolic Regression models
+- Added modified versions of feynman and nguyen benchmarks
+
+**Bug Fixes:**
+
+- Fixed expressions with constants only evaluating to one value instead of an array
 
 
 ### SymbolicRegressionToolkit-1.1.0 (2024-12-10)